progress on fixing duplicate hands bug
