+++ client/components/PlayerHand.js
@@ -1,7 +1,26 @@
 import React from 'react'
 import Card from './Card'
+import {connect} from 'react-redux'
+const playerUpdater = require('../../playerUpdater')
+const handSwap = require('../../handSwap')
 
 class PlayerHand extends React.Component {
+  componentDidUpdate() {
+    if (this.props.readyToPlay === this.props.numPlayers) {
+      this.props.resetPlay()
+      let playerCopy = JSON.parse(JSON.stringify(this.props.me))
+      let updatedPlayer = playerUpdater(
+        playerCopy,
+        this.props.players,
+        this.props.selectedCard
+      )
+      console.log(updatedPlayer)
+      this.props.updatePlayerInStore(updatedPlayer)
+    } else {
+      console.log('NOT READY TO PLAY YET')
+    }
+  }
+
   render() {
     console.log('PlayerHand props: ', this.props)
     return (
@@ -27,4 +46,10 @@ class PlayerHand extends React.Component {
   }
 }
 
-export default PlayerHand
+const mapStateToProps = state => {
+  return {
+    selectedCard: state.selectedCard
+  }
+}
+
+export default connect(mapStateToProps)(PlayerHand)
